+++ app/services/task.py
@@ -1,34 +1,59 @@
+from fastapi import HTTPException, status
 from sqlalchemy.orm import Session
 
-from app.models.task import TaskORM
 from app.repositories.task import TaskRepository
-from app.schemas.task import TaskSchema,TaskCreateSchema,TaskUpdateSchema
+from app.schemas.task import TaskCreateSchema, TaskSchema, TaskUpdateSchema
 
 
 class TaskService:
-    def __init__(self,db:Session)->None:
+    def __init__(self, db: Session) -> None:
         self.db = db
         self.task_repository = TaskRepository(db)
 
-    def list_tasks(self)->list[TaskSchema]:
-        task_orm= self.task_repository.get_all()
+    def list_tasks(self) -> list[TaskSchema]:
+        task_orm = self.task_repository.get_all()
         return [TaskSchema.model_validate(task) for task in task_orm]
 
-    def create_task(self,task_create:TaskCreateSchema)->TaskSchema:
+    def create_task(self, task_create: TaskCreateSchema) -> TaskSchema:
         task_orm = self.task_repository.create(title=task_create.title)
+
         self.db.commit()
+        self.db.refresh(task_orm)
+
         return TaskSchema.model_validate(task_orm)
 
-    def update_task(self,task_id:str,task_update:TaskUpdateSchema)->TaskSchema:
+    def update_task(
+        self,
+        task_id: str,
+        task_update: TaskUpdateSchema,
+    ) -> TaskSchema:
         task_for_update = self.task_repository.get_by_id(task_id=task_id)
+
+        if task_for_update is None:
+            raise HTTPException(
+                status_code=status.HTTP_404_NOT_FOUND,
+                detail="Task not found",
+            )
+
         if task_update.title is not None:
             task_for_update.title = task_update.title
+
         if task_update.completed is not None:
             task_for_update.completed = task_update.completed
+
         self.db.commit()
+        self.db.refresh(task_for_update)
+
         return TaskSchema.model_validate(task_for_update)
 
-    def delete_task(self,task_id:str)->None:
+    def delete_task(self, task_id: str) -> None:
         task_orm = self.task_repository.get_by_id(task_id=task_id)
+
+        if task_orm is None:
+            raise HTTPException(
+                status_code=status.HTTP_404_NOT_FOUND,
+                detail="Task not found",
+            )
+
         self.task_repository.delete(task_orm)
         self.db.commit()
